requirements.txt
accelerate==0.33.0
controlnet-aux==0.0.9
diffusers==0.30.2
einops==0.8.0
fastapi==0.112.2
filelock==3.15.4
gdown==5.2.0
gradio==4.29.0
huggingface-hub==0.24.6
imageio==2.35.1
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.26.1
opencv-python==4.10.0.84
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
timm==0.6.7
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
uvicorn==0.30.6
